likelion13/pyproject.toml
[project]
name = "likelion13"
version = "0.1.0"
description = ""
authors = [
    {name = "chaeyeonlee898",email = "chaeyeonlee898@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "likelion13", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
